requirements.txt
Django==3.2.5
django_crispy_forms==1.14.0
folium==0.12.1.post1
haversine==2.5.1
numpy==1.22.3
pandas==1.4.2
python-dotenv==0.20.0
PyYAML==6.0
whitenoise==6.0.0
gunicorn==20.1.0
user-agents==2.2.0
django-pymemcache==1.0.0
